Create README.md
List of Coding Topics included: 

•	[Topic 1] Servlet.
•	[Topic 2] Collections.
•	[Topic 3] Hibernate.
•	[Topic 4] JSON and HTTP/URL
•	[Topic 5] Threads, Runnables, Executors



I worked for an agroindustry company that needs to know historical weather conditions, daily and forecast. Sometimes the historical condition is difficult to find in the country that I am from, so I created an application for it. I used the API from openweathermap.org to retrieve the weather condition and the forecast for the city of Garzon. To create the historical weather data, I am retrieving the current weather every 15 minutes, and I am storing it on a database. Also, I created an android app where the user can check this data